+++ PROJECTS/02_Dorr/ARK_Lens/tests/tools/build-job-safari-package.js
@@ -1,0 +1,8 @@
+const packageJson = require("../../package.json");
+const { buildJobLensPackage } = require("./job-lens-package");
+
+buildJobLensPackage({
+  manifestSource: "manifests/manifest.job.safari.json",
+  releaseName: `ark-lens-job-search-safari-v${packageJson.version}`,
+  releaseChannel: "automated_safari_staging"
+});